[BUGFIX] Fix set_pet_obj_entry()
set_pte_obj_entry() first creates a new pte containing the address of a KDDM
ObjEntry.

Then, it sets the specific bit pointing that this pte contains an ObjEntry
pointer. At this step, the old pte was used instead of the new one.

At last, it replaces the old pte value with the new one.