[BUGFIX] Fix dup_kddm_obj_entry function.
Port of SVN trunk r5367

The PINNED flag is copied in the new object with all the other
flags. But the freshly allocated object cannot be pinned, so
the PINNED flag has to be cleared.

The bug was leading to random dead-locks in case of an object
shared between several KDDMs.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>