Add a low memory profile in the RPC layer.
Port of SVN trunk r5369 and r5377

In case of high memory pressure, it is now possible to
switch the RPC layer in a low memory consumption mode.
In this mode, the maximum number of buffers usable by
RPCs is much lower.

The 32bit implementation is probably sub-optimal, but less likely to be used
though.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>