[BUGFIX] Fix possible schedule in interrupt.
To wait on a locked page, I was using the kernel
wait_on_page_locked function. But this function
can schedule to wait for the bit to be released.

The wait on the bit has been replaced by an active
test and set on the bit uintil we get the bit.

Since we do not manipulate file linked pages,
pages are not locked for long periods of time.