[BUGFIX] Fix buffer size computation in stat_open.
The buffer size was not correctly computed, leading to
an buffer overflow. The buffer overflow is detected by
the proc_stat kernel function and handled cleanly.

However, in case of buffer overflow, the kernel kfree
the current buffer and kmalloc a new one. When the
kerrighed code try to clean the newly allocated buffer,
it uses a vfree due to the large size of written data.
vfree a kmalloced buffer, is bad !