[BUGFIX] Fix race condition in kcb_zap_pte.
When page_put_kddm_count is called from kcb_zap_pte, the page was not
locked. The atomic dec and test in this function is not enought to
ensure atomicity of future operations. The page can still be mapped
in a one or several processes, even if kddm_count is equal to zero.
Then, the following code can race with init_pte for instance and
setting a freshly allocated object to NULL.

The fix consist in locking the page before calling page_put_kddm_count
to avoid a race with init_pte.