krg_ns: Introduce struct krg_namespace
struct krg_namespace holds all root namespaces of a node that are globalized in
a same cluster. The structure is created whenever a new task is forked wich
creates the required set of namespaces while the cloner has the
GLOBAL_NAMESPACES krgcap. Currently the allowed set is void, so that no
krg_namespace can be created yet. As subsystems will be adapted, the required
set and the allowed set will be extended.

A single such structure is allowed to exist at once, since a node may only
belong to a single cluster at once.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>