[BUGFIX] Fix BUG #92.
If a process is killed during a migration, the remote process can exit
since the original task on the original node is still present in the
system. If we face this case, the mm_struct is removed by the memory
KDDM before the process enter its exit_mm function. Then, the process
exits on a freed mm... and crashes.

This fix introduces a semaphore used to avoid a mm_struct removal in
the KDDM IO memory linker during a process migration.