[FEATURE] Add a new COW break type in kddm_break_cow_object function.
The previous version was breaking a COW by duplicating a page into a
new one and duplicating the corresponding object entry.

During a copy on write request, the page was then copied, sent and
freed. Duplicating the page in this case was just a waste of time.

The new COW break type (KDDM_BREAK_COW_INV), duplicates the object
entry and marks the new one as INVALIDATE. The page is not copied.
It will be used in the copy on write request code to avoid the
useless page copy.