[FEATURE] Add a flush_kddm_event function.
The flush_kddm_event function flush KDDM event pending on
a given object. To do so, we have split the old unique event list
in separate lists stored in the kddm_set struct. When the
flush_kddm_event function is called, we scan this list to find
event matching the requested event and move them from the delayed
wait queue to an active wait queue.