[ENHANCEMENT] Do not reopen file on each VMA import.
Since we have remove the use of DVFS in import/export of
VMA files, we have no more ways to check if a file has
already been imported, to avoid a useless file re-open.

This patch introduces a hashtable used to store files
as they are exported to avoid multiple exports of the same
file. The same way, we check for already imported files
to avoid multiple import and reopen of the same file.

This improves performance of fork06 by 5%.