[TEMPORARY FIX] Limit TIPC hangs on sending side.
Since we currently use one TIPC send for each RPC pack, we
send a lot of TIPC messages for a RPC transaction. This
leads to a rapid overflow of the TIPC send queue which
then blocks the sending queue for a while. All communications
to the corresponding node is stopped for a few ms, which has
a strong impact of performances.

This patch just increases the default send window size. A real
fix would be to merge at least packs from the same RPC into one
unique TIPC message.

By the way, this patch increases fork06 performance by 2% and
increase memory injection by 25%.