[BUGFIX] Fix issue in reconcile_vma for VMA linked to a file.
In case of a process moving back to a previous execution node, the
"ghost" mm remained on the node to back physical pages could have
some VMAs linked to a file. This is typically the case for
initilialized data segment.

This case must be managed. We just have to make sure the incoming
VMA and the one already present on the node points to the same file.