add/procfs: Introduce foreign pidmaps
This commit introduces the concept of foreign pidmaps, which will make it
possible for pids created on a node not living in the current cluster to be
listed by /proc and to be properly marked as allocated whenever the creating
node joins the cluster.

This commit only changes the way procfs lists pids, without any new feature.
Next commit will enhance the introduced infrastructure to support pids not
created on nodes of the current cluster.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>