Injection: Add kddm management in shrink_page_list.
This patch is the main kernel hook of the injection mechanism.
In the kernel page replacement code, the shrink_page_list function is
in charge of writting to disk a set of LRU selected pages.

This patch checks if pages to write are KDDM pages and call the
injection code if needed.