Injection: add KDDM pages management in get_scan_ratio.
get_scan_ratio is in charge of computing the ratio between anonymous
and file pages to swap.

This page adds a third kind of pages in this ratio, the KDDM pages.
KDDM page have a priority ranging from 50% (if there are anonymous
and file pages to swap) to 100% if there are no more other pages
to swap.

By introducing KDDM pages in the ratio of pages to swap, this patch
actually activates the injection code.