Injection: add injection flow control in main kernel swap loop.
Since the RPC layer allocates memory to send messages, we cannot inject
as many pages as we want. Doing so would lead to a very high memory
pressure in a situation where free memory is a rare ressource.

This patch makes sure the RPC layer does not use more than a limited
amount of pages in case of memory injection.