[BUGFIX] Fix SHM incoherence due to bad management of wppage.
The Kerrighed wppage code was never called in case of SYSV memory
segment. shm_vm_ops struct had no wppage function defined.

This patch introduces a wppage function in the shm_vm_ops struct which
calls the kerrighed wppage function stored in krg_shmem_vm_ops, the
same way the existing fault function does.

The only trick is to return EPERM if wppage is not defined in
shmem_vm_ops, which is the case for regular (non Kerrighed) SHM
code. In the do_wp_page function, if we catch a EPERM return from
vm_ops->wppage, we continue with the regular kernel code.
We return a positive error value to avoid conflicts with possible real
errors from effective wppage functions.

A bit tricky but makes it possible to keep as much as possible the
original kernel code and architecture.