[MINOR] Fix location of event_list initialization.
Since event_list can be used a soon as a new KDDM set is stored in the
KDDM hashtable, we must init the event_list in alloc_kddm_set_struct
instead of init_kddm_set.

We also have to init the flags as soon as the structure is accessible
through the hashtable. This field will be used to freeze KDDM during
hotplug operations and then must be initialized as soon as the
hashtable_for_each function can reach a KDDM structure.