[JANITORING] Remove deprecated INV_NO_COPY state.
The INV_NO_COPY state was introduced to distinguish the following 2
cases in kddm_get_object when called with the KDDM_NO_FT_REQ flag:

 - There is no copy on the master node. The local object is then set
   to INV_COPY in the handle_no_object function.

 - There is a copy, we received the copy but it was invalidated
   before the get function was woken up. In this case, the object is
   in state INV_COPY because of the invalidation.

In the first case, we received an answer to our request the get is
done. In the second case, the get is not over since we done not have
the object. These 2 cases have to be distinguished, this is why the
INV_NO_COPY state has been introduced.

The current implementation avoid the invalidation of an object on
which at least one process has a pending get or grab. Thus the second
case is no more possible, making the INV_NO_COPY state useless.