[FEATURE] Add a KDDM_LOCK_FREE flag to the KDDM lookup interface.
This flag will be used to access to objects in the reconfiguration
phase. In this phase, all KDDMs are frozen and any try to access an
object leads to a sleep.

This patch makes it possible to access to objects on frozen KDDM using
the new KDDM_LOCK_FREE flag.