C/R: stronger check before reattaching a semundo to a semaphore array
A semundo is an object storing operations to play to undo some previous
IPC operations on a semaphore array. Such object is related to one
(group) of process(es) and one semaphore array.

When restarting application processes, we are trying to reattach
semundo to the semundo list per semaphore array.

To obtain the semaphore, one used sem_lock. This patch make use of
sem_lock_check instead of sem_lock. sem_lock_check performs stronger
check on the IPC identifier.