[FEATURE] Add a cluster wide barrier mechanism.
The interface of cluster barriers is the following :

* struct cluster_barrier *alloc_cluster_barrier(unique_id_t key)

Allocates a barrier structure for the given barrier key. All nodes
must agree on the barrier key. If no barrier key is given (0), the
function allocates a barrier key. This key must then be sent to other
nodes to allocate local barrier structure.

* void free_cluster_barrier(struct cluster_barrier *barrier)

Free a barrier structure.

* int cluster_barrier(struct cluster_barrier *barrier,
                      krgnodemask_t *nodes,
		      kerrighed_node_t master)

Do a barrier on the barrier structure passed. Nodes involved in the
barrier are passed to the function, as well as the barrier master.
Of course, all nodes must pass the same arguments to this function
when entering a barrier.
The barrier master is the node managing the current ongoing barrier.
This node can be any of the nodes involved in the barrier. It can
change without anything to do from one call of this function to
another.