[BUGFIX] Fix page injection on a dying process.
Between the time a page is selected to be injected by the
shrink_inactive_list function and the time the page is flushed by the
flush_page function, a process holding the page can die and then clear
its anon_vma_kddm_set field.

If so, just ignore and return a swap fail.