[BUGFIX] Fix destroy_kddm_obj_entry concurrency issue.
If the 4 following conditions are fulfilled, we have a corruption of
the prob_owner of the given object:

 * We have a destroy operation on a KDDM object
 * There is a concurrent local get on the same object
 * The prob_owner points to the local node
 * The default owner is not the local node

This patch fixes this issue.

In a destroy operation, if we cannot really destroy the object entry
due to a concurrent access to the object, the prob_owner must be set
to the default owner. Even if we don't really destroy the object, it
is semantically destroyed, thus its prob_owner must be reseted to its
default value.