[BUGFIX] Try to limit RPC memory pressure in injection mechanism.
Since we try to balance the swap of file cache pages and the injection
of KDDM pages, when we inject pages, the page cache can still be quite
big. This patch switches the state of an injecting node to OUT_OF_MEM
as soon as it starts to inject, even if we have a pool of swapable
cache pages.

The main reason for that is to switch the RPC layer to low memory
profile as soon as we have no more free pages, regardless the size of
the page cache.