C/R: handle file substitution at restart time
This patch allows the caller of restart system call to replace any file
(previously open by the application) by a file descriptor of the caller.

The file must be identified by a key (written in file
user_info_<nodeid>.txt).

It can be useful to manage restart of application with sockets, fifos,
etc.