[SWAP-OVER-NFS] Back-port patch to 2.6.30.
Subject: nfs: fix various memory recursions possible with swap over NFS.
Patch-mainline: No
References: FATE#303834
GFP_NOFS is _more_ permissive than GFP_NOIO in that it will initiate IO,
just not of any filesystem data.

The problem is that previuosly NOFS was correct because that avoids
recursion into the NFS code, it now is not, because also IO (swap) can
lead to this recursion.

Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
Signed-off-by: Suresh Jayaraman <sjayaraman@suse.de>