[BUGFIX] IPC: fix crash that happens when cleaning IPC semundos from a bad fork.
When do_fork() fails, namespaces of the child are removed before the
semundos. In Kerrighed kernel, we were trying to access to ipc namespace
to know if the semundos are kerrighed or linux ones. It leads to a
dereferencing of NULL pointer.

This patch uses the ipc namespace of the current task_struct, that is
the parent, in case the child's one is unavailable.