[BUGFIX] Fix BUG 102 - crash in kcb_zap_pte.
In the case of an anonymous file mapped page which has been writen and
then copied on write in an anonymous page, the page index is left to
the old one i.e. the index of the page in the file.

However, since this page is now an anonymous page, it should have as
index value its absolute index, i.e. its index in virtual memory.

This is now fixed in the init_pte function.