[BUGFIX] Fix export of the mm_tasks counter in the import/export of mm_struct.
In case of distant fork, the exported mm is a new one created for the
distant son. However, since the local mm created for the son is a
"ghost" one, i.e. it only contains anonymous mappings to host not yet
migrated pages, counters are not correct. Thus, we must export the
counters of the father BUT the mm_tasks counter. This counter has been
set to 1 for the son, which is the correct value to export.