[BUGFIX] Fix link of VMAs to anonymous KDDM sets.
In check_link_vma_to_anon_memory_kddm_set, checking if a VMA is
anonymous by calling anon_vma is not enough. In fact, this function
checks if a VMA IS anon and not if a VMA COULD BE anon in the future.

The problem mainly comes from anonymous file mappings. Until the first
write to this king of VMA, the corresponding anon_vma field is not set.
To check if this kind of VMA could become anonymous, we check if it is
writable and not shared.

Moreover, we have to check against the shm_vm_ops vm operations struct
to avoid overwritting the SYSV operations.