[JANITORING] Clean-up the reconcile_vmas function.
Since we ensure VMAs are always coherent for distributed mm, we don't
need any more to do complex stuff when we import VMAs.

A VMA is now either not present and must be mapped (this is the case
for shared VMAs and read-only mapped files) either present and must
be identical to the one we receive.