[BUGFIX] Fix issue with page index.
A previous patch was setting the page index of an anonymous page to
the corresponding object id in init_pte, to make sure the index was
correct in any cases.

However, the kernel makes something different for anonymous pages
coming from private mapped files. Even if these pages are anonymous,
the index is the index of the page in the file, not the index of the
page in the process VM. The previous patch was then leading to kernel
BUGs if compiled with DEBUG_VM options. And anyway, it is probably bad
to change the semantic of the page index.

This patch removes the code added in init_pte and modifies the code
where the page index was used in Kerrighed. Since we cannot be sure
the page index corresponds to the index in the VM in any cases, we have
to check the kind of page we are working with and compute the page
index.