[MINOR] Cleanup selection of pages for injection.
Pages was selected for injection by checking the page->obj_entry
field. This was leading to the injection of pages not belonging to a
process with USE_REMOTE_MEMORY capability set (for instance a migrated
process).

Moreover, in case of memory pressure, these pages was moved to the
KDDM lru list. This patch also changes the semantic of this LRU
list. This list does not hold any kind of KDDM pages anymore, but
only KDDM pages with the MIGRATABLE flag set. This flag is set for
pages belonging to a process which has set the USE_REMOTE_MEMORY
capability. Thus "regular" KDDM pages (without the MIGRATABLE flag)
are put in the anon LRU list.