[KDDM] Add an object reference counter in the KDDM object structure.
Since objects can be shared between several KDDM, we need to take
reference counters into account. In the current code, the counts are
managed by the lower layer (KDDM page table for instance). However,
this is not the best place to do that. This patch is the first of a
serie of patches aiming at moving the management of reference counters
in the KDDM layer.