[CLEANUP Cleanup the anon vma KDDM creation code.
This cleanup also integrate the VMA scanning in the KDDM page table
allocation. This avoid some race condition this the swap code.

A mm_struct is marked to be linked to a KDDM as soon as the
mm->anon_vma_kddm_set field is set. In the previous code, the VMA
scanning was done after the setup of this field.