[BUGFIX] krgnodemask.h: Remove abusive BUG_ON
All users of __nth_krgnode() and friends either check that the node mask is not
empty or catch return value KERRIGHED_MAX_NODES.
Moreover it is more efficient to catch KERRIGHED_MAX_NODES than checking
that node mask is not empty. remote_cache_filter.c abusively relies on
this, so make this not abusive anymore.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>