[BUGFIX] Checkpoint: fix race conditions with signal handler asking for exit
Freezing of processes of an application was requested by a signal. Then,
a signal may be posted to request a checkpoint, and processes are woken
up. Before going back to userspace, they handle the signal.

A SIGTERM/SIGKILL signal may be posted after the freeze but BEFORE the
checkpoint signal. In that case, when process is woken up, instead of
handling the checkpoint signal, it first was first handling the exit
signal and then the task disappear. The checkpoint coordinator was then
waiting forever.

This patch fixes this race condition by removing the use of a signal
at checkpoint time.