[BUGFIX] Fix reconcile VMA against anonymous file mapping corner cases.
Fix this function in the following case:
 - new file mmap on a node A with immediate allocation of anon structure.
 - remote mapping of this VMA as an anonymous VMA on a node B.
 - migration of the process to node B.

During the import, we import a file VMA since the local VMA is
anonymous. This is a valid case.

We could remove this case by doing remote mmap for every VMA creation,
including mapped files. This will be mandatory as soon as we will
support threads, but to regular migration and distant forks, this is
useless and avoid costly extra distributed operations.