epm/migration: Make userspace migration requests synchronous
This commit makes userspace migration requests synchronous, that is the
syscall returns once the migration has either completed or failed. On
failure, the error is returned as well.

Note: in the event of several migration attempts racing on the same task
from the same node, errors returned may be mixed between requests.
Adding locking to make this more strict is probably not worth the complexity.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>