[JANITORING] Janitor the KDDM object table and object locking.
Since we have introduced a way to write an external object table
mechanism, the original table locking is deprecated and can lead to 
locking issues.

The new locking system is based on:
 - An object lock, implemented as a bit in the object flags
   field. There is no major changes in this lock. This lock is
   equivalent to a spinlock. It protects the kddm_obj structure.

 - A table lock, implemented as a function pointer defined by the
   object table code. This code is a new one introduced in this
   patch. The table protection depends on the table and is implemented
   by the table code designer. This lock protects the table during
   object operations on the table like add, lookup and remove.

 - An object frozen flag, implemented as a counter in the object
   structure. There is no changes in this code. The frozen flag
   protects the object from some protocol operations (copy on write
   request, invalidation request, etc). It is mainly used to protect
   an object between a get or grab and a put.
   In this new object locking mechanism, we use the frozen flag to
   avoid any modification of the object table slot: the object cannot
   be COW or removed from the table.

In this new implementation, table operations are now really protected
by a locking system defined by the table operations, no more by using
the generic KDDM mutexes.

The second change is the duration the table lock is taken. For most
operations, the lock is just taken during the object lookup, following
this scheme:

 - Object lookup
    - lock the table
    - lookup the object in the table
    - lock the object
    - unlock the table
 - Use the object
 - Unlock the object

In some situations, we need to release the object lock (to allow a
sleep for instance) but don't want the object to be modified. In this
case, we freeze the object and release the lock. Since the object is
frozen, only "minor" operations are allowed on the object (local get,
some ACK reception, etc). By freezing the object, we are sure the
obj_entry structure we have will still be valid without requiering a
new table lookup.

 - Object lookup
    - lock the table
    - lookup the object in the table
    - lock the object
    - unlock the table
 - Use the object
 - freeze the object
 - Unlock the object
 - Do might_sleep operation
 - Lock the object
 - Unfreeze the object
 - Continue using the object
 - Unlock the object