[BUGFIX] Add KDDM_OWNER_OBJ flag to the WAIT_CHG_OWN_ACK state.
An object in state WAIT_CHG_OWN_ACK is owner of the object until we
receive an ACK from the node we have transfer the ownsership to.

Thus, the WAIT_CHG_OWN_ACK state must have the KDDM_OWNER_OBJ flag.
It's even more important since we have modified the
change_object_state function to change the prob owner depending on the
KDDM_OWNER_OBJ flag.