hotplug/remove: Use the hotplug coordinator
Since the node requesting the remove operation may be removed by this
operation and must contact the coordinator through
hotplug_finish_request() at the end of the operation, after acknowledging
the requesting node all nodes wait for a last acknowledgement before
completing the operation. This will be mandatory when dynamic RPC
connections will be implemented to support rebooting nodes.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>