dvfs: Don't pass freed file to end_import_dvfs_file()
import_one_open_file() gives a freed file to end_import_dvfs_file()
whenever a file descriptor is already attached to the dvfs_file. This does
not harm since first_import == 0 in this case and thus end_import_dvfs_file()
does not use the passed file descriptor, but later evolutions of
end_import_dvfs_file() could take the file descriptor as valid, so ensure
that this is the case.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>