[BUGFIX] Fix a LRU list corruption with SHM pages.
This is not a real fix but a work-around. To avoid LRU list issues, we
prevent the SHM pages to be inserted in such lists ! Since we
currently do not support the swap of shm pages, this is not such an
issue. But we will have to do something on this topic.