[BUGFIX] Fix crash in destroy_kddm_obj_entry: BUG_ON(object_frozen(...))
In the case of a call to _kddm_remove_frozen_object, the object must
have been grabbed before to ensure cluster wide atomicity of the
remove and concurrent accesses to the object.

Grabbing the object using _kddm_grab_object_no_ft is not enough in
some situations. For instance in the case of an object which does not
exist and with a remote owner, the object will not be grabbed. The
local remove can then conflict with a remove acces to the object.

A call to _kddm_grab_object_manual_ft ensures the object will be
grabbed, even if the object does not exists. Actually, the ownership
of this object will be moved to the grabbing node.