krgsched: Introduce ConfigFS entries for remote values in probe sources
This patch is intended to modify the behavior of SchedConfig's view
through ConfigFS in order to expose the remote nodes probe's sources
values.

That is, before we had something like:
/config/krg_schedulers/probes/probe_name/probe_source_name/value.
Now, we also provide path like:
/config/krg_schedulers/probes/probe_name/probe_source_name/nodeid/value
for each node (nodeid) in the cluster.

This is needed for distributed scheduling done in userspace (thanks to
the migrate(2) syscall).

The functionnality is implemented using RPC, and by creating a
configfs-modified data structure when calling
scheduler_probe_source_create(). We copy the original
configfs_attributes and then we modify its properties to :
 * Display the node id of the targetted node
 * Make use of our specific configfs_item_operations

This last one references the
scheduler_probe_source_attribute_show_remote function, which is in
charge of creating, sending and waiting for the RPC request response.
The RPC request is handled by handle_scheduler_pipe_show_remote_value
and simply read local value of the probe source, and returning it.

[ Subject edited for clearer description, and error handling fixup and
  cleanup -- Louis Rilling ]

Signed-off-by: Alexandre Lissy <alexandre.lissy@etu.univ-tours.fr>
Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>